FC-365 rounding to 2 dp on National Average 2 report.

+++ JsPlc.Ssc.PetrolPricing/JsPlc.Ssc.PetrolPricing.Models/ViewModels/NationalAverageReportBrandViewModel.cs
@@ -5,10 +5,10 @@ namespace JsPlc.Ssc.PetrolPricing.Models.ViewModels
     {
         public string BrandName { get; set; }
         
-        public int Average { get; set; }
+        public double Average { get; set; }
 
-        public int Min { get; set; }
+        public double Min { get; set; }
 
-        public int Max { get; set; }
+        public double Max { get; set; }
     }
 }
